requirements.txt
   pyspark==3.4.0
   mysql-connector-python==8.0.33
   pyiceberg==0.4.0
   boto3==1.26.137
   python-dotenv==1.0.0